On-the-fly compilation (#55)
* added sources

* renaming

* deleted test folder

* edited Manifest.in

* added compilation on non-Windows

* edited paths check

* fixed links to parameter tuning guide

* hotfix

* added compilation on Windows

* edited appveyor

* added MSBuild compilation on Windows

* hothix

* hothix

* fixed error when module hasn't been loaded

* hotfix

* added Windows x64 platforms to appveyor

* fixed x86 configuration

* edited success compilation of executable

* added compilation on Windows

* Windows compatibility

* added more compilation variants

* fixed makefile for MinGW

* added preprocessor definition

* hotfix

* fixed appveyor PATH

* fixed g++ crashes on Windows x86

* fixed caps for pretty looking

* fixed compiling on Windows with g++

* edited existing VS solution

* edited _is_executable_response for 100% sure check

* editet Manifest.py

* fixed logging

* fixed pretty looking of log

* fixed deleting temp files created while checking exe file

* edited Readme

* fiixed Readme

* fixed Readme

* fixed typos in Readme

* fixed link in Readme

* edited Readme

* fixed Readme

* added precompiled exe for Win32

* fixed solution files

* deleted unnecessary import

* improved clear_folder

* updated CI

* updated scikit-learn 0.19 compatibility

* fixed typo

* deleted license

* Delete LICENSE

* Create LICENSE

* fixed license references

+++ rgf/sklearn.py
@@ -72,20 +72,59 @@ def _get_paths():
     return def_exe, exe, temp
 
 
+_DEFAULT_EXE_PATH, _EXE_PATH, _TEMP_PATH = _get_paths()
+
+
+if not os.path.isdir(_TEMP_PATH):
+    os.makedirs(_TEMP_PATH)
+if not os.access(_TEMP_PATH, os.W_OK):
+    raise Exception("{0} is not writable directory. Please set "
+                    "config flag 'temp_location' to writable directory".format(_TEMP_PATH))
+
+
 def _is_executable_response(path):
+    temp_x_loc = os.path.join(_TEMP_PATH, 'temp.train.data.x')
+    temp_y_loc = os.path.join(_TEMP_PATH, 'temp.train.data.y')
+    np.savetxt(temp_x_loc, [[1, 0, 1, 0], [0, 1, 0, 1]], delimiter=' ', fmt="%s")
+    np.savetxt(temp_y_loc, [1, -1], delimiter=' ', fmt="%s")
+    _UUIDS.append('temp')
+    params = []
+    params.append("train_x_fn=%s" % temp_x_loc)
+    params.append("train_y_fn=%s" % temp_y_loc)
+    params.append("model_fn_prefix=%s" % os.path.join(_TEMP_PATH, "temp.model"))
+    params.append("reg_L2=%s" % 1)
+
     try:
-        subprocess.check_output((path, "train"))
+        subprocess.check_output((path, "train", ",".join(params)))
         return True
     except Exception:
         return False
 
 
+if _is_executable_response(_DEFAULT_EXE_PATH):
+    _EXE_PATH = _DEFAULT_EXE_PATH
+elif _is_executable_response(os.path.join(os.path.dirname(__file__), _DEFAULT_EXE_PATH)):
+    _EXE_PATH = os.path.join(os.path.dirname(__file__), _DEFAULT_EXE_PATH)
+elif not os.path.isfile(_EXE_PATH):
+    raise Exception("{0} is not executable file. Please set "
+                    "config flag 'exe_location' to RGF execution file.".format(_EXE_PATH))
+elif not os.access(_EXE_PATH, os.X_OK):
+    raise Exception("{0} cannot be accessed. Please set "
+                    "config flag 'exe_location' to RGF execution file.".format(_EXE_PATH))
+elif _is_executable_response(_EXE_PATH):
+    pass
+else:
+    raise Exception("{0} does not exist or {1} is not in the "
+                    "'PATH' variable.".format(_EXE_PATH, _DEFAULT_EXE_PATH))
+
+
 @atexit.register
 def _cleanup():
-    for uuid in _UUIDS:
-        model_glob = os.path.join(_TEMP_PATH, uuid + "*")
-        for fn in glob(model_glob):
-            os.remove(fn)
+    if _UUIDS is not None:
+        for uuid in _UUIDS:
+            model_glob = os.path.join(_TEMP_PATH, uuid + "*")
+            for fn in glob(model_glob):
+                os.remove(fn)
 
 
 def _sigmoid(x):
@@ -229,28 +268,6 @@ class _AtomicCounter(object):
 
 
 _COUNTER = _AtomicCounter()
-_DEFAULT_EXE_PATH, _EXE_PATH, _TEMP_PATH = _get_paths()
-
-if _is_executable_response(_DEFAULT_EXE_PATH):
-    _EXE_PATH = _DEFAULT_EXE_PATH
-elif not os.path.isfile(_EXE_PATH):
-    raise Exception("{0} is not executable file. Please set "
-                    "config flag 'exe_location' to RGF execution file.".format(_EXE_PATH))
-elif not os.access(_EXE_PATH, os.X_OK):
-    raise Exception("{0} cannot be accessed. Please set "
-                    "config flag 'exe_location' to RGF execution file.".format(_EXE_PATH))
-elif _is_executable_response(_EXE_PATH):
-    pass
-else:
-    raise Exception("{0} does not exist or {1} is not in the "
-                    "'PATH' variable.".format(_EXE_PATH, _DEFAULT_EXE_PATH))
-
-if not os.path.isdir(_TEMP_PATH):
-    os.makedirs(_TEMP_PATH)
-if not os.access(_TEMP_PATH, os.W_OK):
-    raise Exception("{0} is not writable directory. Please set "
-                    "config flag 'temp_location' to writable directory".format(_TEMP_PATH))
-
 
 
 class RGFClassifier(BaseEstimator, ClassifierMixin):
@@ -258,7 +275,7 @@ class RGFClassifier(BaseEstimator, ClassifierMixin):
     A Regularized Greedy Forest [1] classifier.
 
     Tuning parameters detailed instruction:
-        http://tongzhang-ml.org/software/rgf/rgf1.2-guide.pdf
+        https://github.com/fukatani/rgf_python/blob/master/include/rgf/rgf1.2-guide.pdf
 
     Parameters
     ----------
@@ -693,7 +710,7 @@ class RGFRegressor(BaseEstimator, RegressorMixin):
     A Regularized Greedy Forest [1] regressor.
 
     Tuning parameters detailed instruction:
-        http://tongzhang-ml.org/software/rgf/rgf1.2-guide.pdf
+        https://github.com/fukatani/rgf_python/blob/master/include/rgf/rgf1.2-guide.pdf
 
     Parameters
     ----------
